#6: Moved SyftController to the Network package, which is also re-organized.
