+++ Assets/Network/Utils.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: e6657a556ad154c7bae69a915b6f8c2a
+folderAsset: yes
+timeCreated: 1510597759
+licenseType: Free
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
